refactor(tests): split fake repository projects
